blueprint: Rename tester.js to jenkins.js
An upcoming commit will introduce another service to the tester
deployment.
